refactor(schemas): migrate Phase 15 endpoint schemas — 5 files, 22 classes (#6042)
- database_mcp.py: migrate MCPTool→DatabaseMCPTool, SQLQueryRequest, SQLExecuteRequest,
  SchemaRequest, TableListRequest → schemas_code.py (preserves field_validators)
- integration_project_management.py: migrate ConnectionTestRequest, ProviderInfo,
  IssueCreateRequest, IssueUpdateRequest, SearchRequest→ProjectMgmtSearchRequest
  → schemas_workflows.py
- knowledge_relations.py: migrate CreateRelationRequest, DeleteRelationRequest,
  TraverseRequest, HybridSearchRequest → schemas_knowledge.py
- nl_database.py: migrate NLQueryRequest, TrainRequest, NLQueryResponse,
  TrainResponse → schemas_knowledge.py
- merge_conflict_resolution.py: migrate ConflictAnalysisRequest, ConflictResolutionRequest,
  RepositoryAnalysisRequest, ApplyResolutionRequest → schemas_code.py

Renames to avoid name collisions:
- MCPTool → DatabaseMCPTool (collides with prometheus MCPTool, both now distinct)
- SearchRequest → ProjectMgmtSearchRequest (collides with knowledge SearchRequest)

Cleanup: removed dead-code SearchRequest import in integration_project_management.py
(class was defined but search_issues endpoint uses Query parameters, not the model).

Constants: schemas_code.py adds _DB_MCP_MAX_RESULT_ROWS=1000 and
_DB_MCP_ALLOWED_DML_OPERATIONS for SQL validators.
